Add files via upload
These visualizations come from the 2023 Missouri DataFest competition. The data comes from a company providing pro-bono legal services via call ins. The data housed a number of different predictors, but our group focused on the transcribed conversations that actually took place between legal experts and those seeking help. We then took those conversations and analyzed them with a natural language processor that would input a value from 1 to -1 based on the "sentimentality" (tone) of the conversation. The geographical visualizations were done in Tableau, and the boxplot was done in R.