Add files via upload
This plot contrasts the number of cars in the traffic dataset that coordinates with a "good" or "bad" traffic condition. The labels describe a binomial outcome of traffic data that was determined a computer vision model. I was testing whether the computer vision model determined its traffic conditions solely based off of the number of cars that passed through the given location. If this was the case, then I wouldn't have anything to model. Further information is provided in my Traffic Conditions Project Repo.