Add files via upload
This is a screenshot of a dashboard I created on the 2021 version of the annual business survey. The data is gathered by the Census Bureau as a survey of all companies in the country. It gives summary information about the employers on an industry basis (how many businesses in the construction sector were owned by white, nonveteran, females for example). My dashboard allows the user to select any industry or combination of industries and view the ownership numbers across race, sex, and veteran status. Contact me if you want to test the actual dashboard.